(WIP) Old Blender mock things (see extended message)
We were originally trying to mock Blender's bpy module for automated testing. We
decided to take a different approach with running the entire test suite in
Blender itself. Some of this code is still useful for that. Some of it isn't,
but we're keeping it around for reference purposes. Also we did spend a lot of
time on it and might revisit it at some point, maybe.